feat: change Navbar color
