+++ src/components/HomeScreen/Navbar/Navbar.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from "@testing-library/react";
-import MenuComponent from "./MenuComponent";
+import Navbar from "./Navbar";
 import { describe, test, expect } from "vitest";
 
 // TODO: fix file
-// describe("<MenuComponent />", () => {
+// describe("<Navbar />", () => {
 //   test("should render About author -button and show more -button", () => {
 //     // different screen sizes render different buttons,
 //     // but test does not recognize this
@@ -17,7 +17,7 @@ import { describe, test, expect } from "vitest";
 //     // setIsLoggedIn: React.Dispatch<React.SetStateAction<boolean>>;
 //     // setOpenCreateAccount: React.Dispatch<React.SetStateAction<boolean>>;
 
-//     render(<MenuComponent setOpenAbout={setOpenAbout} />);
+//     render(<Navbar setOpenAbout={setOpenAbout} />);
 
 //     const element = screen.getByRole("button", {
 //       name: "about-author-big",
